requirements.txt
requests==2.26.0
python-dotenv==0.19.0 